Removed ..dtframe directive (no longer used)
