+++ docs/conf.py
@@ -49,7 +49,6 @@ needs_sphinx = '1.8'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    '_ext.dtframe_directive',
     '_ext.xcode',
     '_ext.xcontributors',
     '_ext.xfunction',
